VB-5218, refactor clear notifcations
